[@types/ember] add ember__template (#38571)
* [@types/ember] add ember__template

* [@types/ember] move -private/handlebars into @ember/template

* [@types/ember] fix SafeString namespacing
