+++ types/ember__string/ember__string-tests.ts
@@ -1,5 +1,4 @@
-import { dasherize, camelize, capitalize, classify, decamelize, htmlSafe, loc, underscore, w, isHTMLSafe } from '@ember/string';
-import { SafeString } from '@ember/string/-private/handlebars';
+import { dasherize, camelize, capitalize, classify, decamelize, loc, underscore, w } from '@ember/string';
 
 dasherize(); // $ExpectError
 dasherize('blue man group'); // $ExpectType string
@@ -32,15 +31,3 @@ capitalize('', ''); // $ExpectError
 loc(); // $ExpectError
 loc("_Hello World");  // $ExpectType string
 loc("_Hello %@ %@", ["John", "Smith"]);  // $ExpectType string
-
-const handlebarsSafeString: SafeString = htmlSafe('lorem ipsum...');
-htmlSafe('lorem ipsum...'); // $ExpectType SafeString
-const regularString: string = htmlSafe('lorem ipsum...'); // $ExpectError
-
-function isSafeTest(a: string | SafeString) {
-  if (isHTMLSafe(a)) {
-      a = a.toString();
-  }
-
-  camelize(a);
-}
