+++ types/ember__string/index.d.ts
@@ -4,15 +4,13 @@
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 // TypeScript Version: 2.8
 
-import { SafeString } from "./-private/handlebars";
+export { htmlSafe, isHTMLSafe } from '@ember/template';
 
 export function camelize(str: string): string;
 export function capitalize(str: string): string;
 export function classify(str: string): string;
 export function dasherize(str: string): string;
 export function decamelize(str: string): string;
-export function htmlSafe(str: string): SafeString;
-export function isHTMLSafe(str: any): str is SafeString;
 export function loc(template: string, args?: string[]): string;
 export function underscore(str: string): string;
 export function w(str: string): string[];
